Delete README.md
Organizational Chart App - MEAN Stack App
-----------------------------------------

The organization chart is a diagram showing graphically the relation of one official to another, or others, of a company. It is also used to show the relation of one department to another, or others, or of one function of an organization to another, or others. This chart is valuable in that it enables one to visualize a complete organization, by means of the picture it presents.

A company's organizational chart typically illustrates relations between people within an organization. Such relations might include managers to sub-workers, directors to managing directors, chief executive officer to various departments, and so forth. 

Instalation notes
-----------------
1. Open console in project's folder
1. Type "npm install" in the console
2. Type "bower install -d" in the console
3. Type "node server.js" in the console
4. Open "http://localhost:88" in web browser